+++ src/main/java/roomescape/global/error/exception/BadRequestException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.global.error.exception;
 
 public class BadRequestException extends IllegalArgumentException {
     public BadRequestException(String message) {
