+++ src/main/java/roomescape/global/error/exception/ConflictException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.global.error.exception;
 
 public class ConflictException extends IllegalArgumentException {
     public ConflictException(String message) {
