+++ src/main/java/roomescape/global/error/exception/NotFoundException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.global.error.exception;
 
 public class NotFoundException extends IllegalArgumentException {
     public NotFoundException(String message) {
